requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
h5py>=3.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
jupyter>=1.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0
pytest>=7.4.0
black>=23.0.0
yfinance>=0.2.0